initial event and logic demo is completed
add a conway game of life layer in there to test out the extensibility of the layering system.

now i've also got a way to somewhat visual what is going on and to step through the frames of the game.

Got a little bit of visibility into the eventing system so i should be able to find out if there is anything particularly really weird going on while moving forward.